shifted some functions around to tidy up repo
