+++ src/arc_tigers/training/utils.py
@@ -1,31 +1,10 @@
 from collections import Counter
 
-import numpy as np
 import torch
-from datasets import Dataset, concatenate_datasets, load_dataset
 from torch import nn
 from transformers import Trainer
 
-from arc_tigers.constants import DATA_DIR
-from arc_tigers.data.utils import (
-    BINARY_COMBINATIONS,
-    ONE_VS_ALL_COMBINATIONS,
-    get_target_mapping,
-    preprocess_function,
-)
 from arc_tigers.utils import get_device
-
-
-def balance_dataset(dataset):
-    # Get the number of samples in each class
-    label_counts = Counter(dataset["label"])
-    # Calculate the number for each class
-    min_samples = min(label_counts.values())
-    resampled_data = []
-    for label in label_counts:
-        label_data = dataset.filter(lambda x, label=label: x["label"] == label)
-        resampled_data.append(label_data.shuffle().select(range(min_samples)))
-    return concatenate_datasets(resampled_data)
 
 
 class WeightedLossTrainer(Trainer):
@@ -47,98 +26,6 @@ class WeightedLossTrainer(Trainer):
         return (loss, outputs) if return_outputs else loss
 
 
-def get_reddit_data(
-    setting, target_config, balanced, n_rows, tokenizer, random_seed=42
-):
-    """
-    Loads and preprocesses the Reddit dataset based on the specified configuration.
-
-    Args:
-        setting (str): The classification setting, either "multi-class" or "one-vs-all".
-        target_config (str): The target configuration to use for the dataset.
-        balanced (bool): Whether to balance the dataset by resampling classes.
-        n_rows (int): The number of rows to load from the dataset.
-        tokenizer (transformers.PreTrainedTokenizer): The tokenizer to use for
-        reprocessing.
-
-    Raises:
-        ValueError: If an unknown setting is provided.
-
-    Returns:
-        tuple: A tuple containing the training dataset, evaluation dataset, and test
-        dataset.
-    """
-    split_generator = np.random.default_rng(seed=random_seed)
-    # work out the data directory
-    data_dir = f"{DATA_DIR}/reddit_dataset_12/{n_rows}_rows/splits/{target_config}/"
-    # load dataset
-    dataset: dict[str, Dataset] = load_dataset(
-        "csv",
-        data_files={"train": f"{data_dir}/train.csv", "eval": f"{data_dir}/test.csv"},
-    )
-    meta_data = {}
-
-    if setting == "multi-class":
-        train_targets = BINARY_COMBINATIONS[target_config]["train"]
-        train_dataset = dataset["train"].filter(
-            lambda y: y in train_targets, input_columns=["label"]
-        )
-        train_target_map = get_target_mapping(setting, train_targets)
-        meta_data["train_target_map"] = train_target_map
-
-    elif setting == "one-vs-all":
-        train_dataset = dataset["train"]
-        train_targets = ONE_VS_ALL_COMBINATIONS[target_config]["train"]
-        test_targets = ONE_VS_ALL_COMBINATIONS[target_config]["test"]
-        train_target_map = get_target_mapping(setting, train_targets)
-        test_target_map = get_target_mapping(setting, test_targets)
-        meta_data["train_target_map"] = train_target_map
-        meta_data["test_target_map"] = test_target_map
-
-    else:
-        err_msg = (
-            f"Unknown setting: {setting}. Please use 'multi-class' or 'one-vs-all'."
-        )
-        raise ValueError(err_msg)
-
-    tokenized_train_dataset = train_dataset.map(
-        preprocess_function,
-        batched=True,
-        fn_kwargs={
-            "tokenizer": tokenizer,
-            "targets": train_target_map,
-        },
-    )
-    if setting == "one-vs-all":
-        tokenized_test_dataset: Dataset = dataset["test"].map(
-            preprocess_function,
-            batched=True,
-            fn_kwargs={
-                "tokenizer": tokenizer,
-                "targets": test_target_map,
-            },
-        )
-    elif setting == "multi-class":
-        # use the train set so hat the classes are consistent
-        tokenized_train_dataset, tokenized_test_dataset = (
-            tokenized_train_dataset.train_test_split(
-                test_size=0.5, generator=split_generator
-            ).values()
-        )
-
-    # balance the dataset
-    if balanced:
-        print("Balancing the datasets...")
-        tokenized_train_dataset = balance_dataset(tokenized_train_dataset)
-        tokenized_test_dataset = balance_dataset(tokenized_test_dataset)
-
-    # Split dataset
-    train_data, eval_data = tokenized_train_dataset.train_test_split(
-        test_size=0.1, generator=split_generator
-    ).values()
-    return train_data, eval_data, tokenized_test_dataset, meta_data
-
-
 def get_label_weights(dataset, verbose=True):
     label_counter = Counter(dataset["label"])
     label_counts = dict(sorted(label_counter.items(), key=lambda item: item[1]))
@@ -152,4 +39,4 @@ def get_label_weights(dataset, verbose=True):
         for label_idx, (label, count) in enumerate(label_counts.items()):
             print(f"Label: {label}, Count: {count}, Weight: {label_weights[label_idx]}")
 
-    return label_weights
+    return label_weights, label_counts
